답글 있는 원댓글 삭제 구현
- 답글 있는 원댓글 삭제 시 "삭제된 댓글입니다"라고 표시
- 그러나 나머지 답글까지 삭제하면 삭제된 댓글이라는 표시가 그대로 남아 있어서 수정해야 됨...

+++ models/postModel.js
@@ -222,8 +222,19 @@ const commentModel = {
   },
 
   deleteComments: (cmt_num, callback) => {
-    db.query('DELETE FROM comment WHERE cmt_num = ?', [cmt_num], () => {
-      callback();
+    db.query('SELECT COUNT(*) as count FROM comment WHERE cmt_refnum = ?', [cmt_num], (error, results) => {
+        if (error) {
+            throw error;
+        }
+        if (results[0].count == 0) {
+            db.query('DELETE FROM comment WHERE cmt_num = ?', [cmt_num], () => {
+                callback();
+            });
+        } else {
+            db.query('UPDATE comment SET is_deleted = 1 WHERE cmt_num = ?', [cmt_num], () => {
+                callback();
+            });
+        }
     });
   },
